requirements.txt
Flask==1.0.2
gunicorn==19.9.0
jsonify
mysql-connector-python==8.0.13
